plugin_controller/requirements.txt
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
decorator==4.3.0
dict-to-protobuf==0.0.3.8
docker==3.2.1
docker-pycreds==0.2.2
idna==2.6
more-itertools==4.1.0
pluggy==0.6.0
protobuf==3.6.0
py==1.5.3
pytest==3.5.0
pytest-cov==2.5.1
python-magic==0.4.15
ramrodbrain>=0.1.31
requests==2.18.4
rethinkdb==2.3.0.post6
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
